+++ today-photo/today-photo/UserAuthModel.swift
@@ -17,28 +17,61 @@ class UserAuthModel: ObservableObject {
     @Published var errorMessage: String = ""
     
     init() {
-        checkStatus()
+        check()
     }
     
     func checkStatus() {
-        GIDSignIn.sharedInstance.restorePreviousSignIn { user, error in
-            guard error == nil else { return }
+        if GIDSignIn.sharedInstance.currentUser != nil {
+            let user = GIDSignIn.sharedInstance.currentUser
+            guard let user = user else { return }
             
-            if GIDSignIn.sharedInstance.currentUser != nil {
-                let user = GIDSignIn.sharedInstance.currentUser
-                guard let user = user else { return }
-                
-                let givenName = user.profile?.givenName
-                let profilePicUrl = user.profile!.imageURL(withDimension: 100)!.absoluteString
-                
-                self.givenName = givenName ?? ""
-                self.profilePicUrl = profilePicUrl
-                self.isLoggedIn = true
-            } else {
-                self.givenName = ""
-                self.profilePicUrl = ""
-                self.isLoggedIn = false
-            }
+            let givenName = user.profile?.givenName
+            let profilePicUrl = user.profile!.imageURL(withDimension: 100)!.absoluteString
+            
+            self.givenName = givenName ?? ""
+            self.profilePicUrl = profilePicUrl
+            self.isLoggedIn = true
+        } else {
+            self.givenName = ""
+            self.profilePicUrl = ""
+            self.isLoggedIn = false
         }
     }
+    
+    func check() {
+        GIDSignIn.sharedInstance.restorePreviousSignIn { _, error in
+            guard error == nil else { return }
+        }
+        self.checkStatus()
+    }
+    
+    func signIn() {
+        guard let presentingViewController = (UIApplication.shared.connectedScenes.first as? UIWindowScene)?.windows.first?.rootViewController else {return}
+        guard let clientID = FirebaseApp.app()?.options.clientID else { return }
+        let signInConfig = GIDConfiguration(clientID: clientID)
+        
+        GIDSignIn.sharedInstance.signIn(
+            with: signInConfig,
+            presenting: presentingViewController) { user, error in
+                guard error == nil else { return }
+                
+                guard
+                    let authentication = user?.authentication,
+                    let idToken = authentication.idToken
+                else { return }
+                
+                let credential = GoogleAuthProvider.credential(withIDToken: idToken, accessToken: authentication.accessToken)
+                
+                Auth.auth().signIn(with: credential) { _, _ in
+                // If sign in succeeded, display the app's main content View.
+                }
+                self.checkStatus()
+            }
+    }
+    
+    func signOut() {
+        GIDSignIn.sharedInstance.signOut()
+        self.checkStatus()
+    }
+    
 }
