--- today-photo/today-photo/View/SettingView.swift
@@ -1,20 +1,0 @@
-//
-//  SettingView.swift
-//  today-photo
-//
-//  Created by 조희라 on 2022/11/26.
-//
-
-import SwiftUI
-
-struct SettingView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct SettingView_Previews: PreviewProvider {
-    static var previews: some View {
-        SettingView()
-    }
-}
