+++ today-photo/today-photo/View/TestView.swift
@@ -2,14 +2,25 @@
 //  TestView.swift
 //  today-photo
 //
-//  Created by 조희라 on 2022/11/23.
+//  Created by 조희라 on 2022/11/26.
 //
 
 import SwiftUI
 
 struct TestView: View {
+    @EnvironmentObject var userVM: UserAuthModel
+    
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        AsyncImage(url: URL(string: userVM.profilePicUrl))
+        
+        Text(userVM.givenName)
+        
+        Button {
+            userVM.signOut()
+        } label: {
+            Text("로그아웃")
+        }
+
     }
 }
